Refactor out concurrency primitives and clean out debug info
