+++ src/jitlayers.h
@@ -15,9 +15,6 @@
 #include "julia_assert.h"
 #include "debug-registry.h"
 
-#include <stack>
-#include <queue>
-
 // As of LLVM 13, there are two runtime JIT linker implementations, the older
 // RuntimeDyld (used via orc::RTDyldObjectLinkingLayer) and the newer JITLink
 // (used via orc::ObjectLinkingLayer).
@@ -64,33 +61,7 @@ static inline bool imaging_default() {
     return jl_options.image_codegen || (jl_generating_output() && !jl_options.incremental);
 }
 
-struct jl_locked_stream {
-    JL_STREAM *stream = nullptr;
-    std::mutex mutex;
-
-    struct lock {
-        std::unique_lock<std::mutex> lck;
-        JL_STREAM *&stream;
-
-        lock(std::mutex &mutex, JL_STREAM *&stream) : lck(mutex), stream(stream) {}
-
-        JL_STREAM *&operator*() {
-            return stream;
-        }
-
-        explicit operator bool() {
-            return !!stream;
-        }
-
-        operator JL_STREAM *() {
-            return stream;
-        }
-    };
-
-    lock operator*() {
-        return lock(mutex, stream);
-    }
-};
+typedef jl_cc::Locked<JL_STREAM*> jl_locked_stream;
 
 typedef struct _jl_llvm_functions_t {
     std::string functionObject;     // jlcall llvm Function name
@@ -225,111 +196,6 @@ public:
     typedef orc::IRCompileLayer CompileLayerT;
     typedef orc::IRTransformLayer OptimizeLayerT;
     typedef object::OwningBinary<object::ObjectFile> OwningObj;
-    template
-    <typename ResourceT, size_t max = 0,
-        typename BackingT = std::stack<ResourceT,
-            std::conditional_t<max == 0,
-                SmallVector<ResourceT>,
-                SmallVector<ResourceT, max>
-            >
-        >
-    >
-    struct ResourcePool {
-        public:
-        ResourcePool(std::function<ResourceT()> creator) : creator(std::move(creator)), mutex(std::make_unique<WNMutex>()) {}
-        class OwningResource {
-            public:
-            OwningResource(ResourcePool &pool, ResourceT resource) : pool(pool), resource(std::move(resource)) {}
-            OwningResource(const OwningResource &) = delete;
-            OwningResource &operator=(const OwningResource &) = delete;
-            OwningResource(OwningResource &&) = default;
-            OwningResource &operator=(OwningResource &&) = default;
-            ~OwningResource() {
-                if (resource) pool.release(std::move(*resource));
-            }
-            ResourceT release() {
-                ResourceT res(std::move(*resource));
-                resource.reset();
-                return res;
-            }
-            void reset(ResourceT res) {
-                *resource = std::move(res);
-            }
-            ResourceT &operator*() {
-                return *resource;
-            }
-            ResourceT *operator->() {
-                return get();
-            }
-            ResourceT *get() {
-                return resource.getPointer();
-            }
-            const ResourceT &operator*() const {
-                return *resource;
-            }
-            const ResourceT *operator->() const {
-                return get();
-            }
-            const ResourceT *get() const {
-                return resource.getPointer();
-            }
-            explicit operator bool() const {
-                return resource;
-            }
-            private:
-            ResourcePool &pool;
-            llvm::Optional<ResourceT> resource;
-        };
-
-        OwningResource operator*() {
-            return OwningResource(*this, acquire());
-        }
-
-        OwningResource get() {
-            return **this;
-        }
-
-        ResourceT acquire() {
-            std::unique_lock<std::mutex> lock(mutex->mutex);
-            if (!pool.empty()) {
-                return pop(pool);
-            }
-            if (!max || created < max) {
-                created++;
-                return creator();
-            }
-            mutex->empty.wait(lock, [&](){ return !pool.empty(); });
-            assert(!pool.empty() && "Expected resource pool to have a value!");
-            return pop(pool);
-        }
-        void release(ResourceT &&resource) {
-            std::lock_guard<std::mutex> lock(mutex->mutex);
-            pool.push(std::move(resource));
-            mutex->empty.notify_one();
-        }
-        private:
-        template<typename T, typename Container>
-        static ResourceT pop(std::queue<T, Container> &pool) {
-            ResourceT top = std::move(pool.front());
-            pool.pop();
-            return top;
-        }
-        template<typename PoolT>
-        static ResourceT pop(PoolT &pool) {
-            ResourceT top = std::move(pool.top());
-            pool.pop();
-            return top;
-        }
-        std::function<ResourceT()> creator;
-        size_t created = 0;
-        BackingT pool;
-        struct WNMutex {
-            std::mutex mutex;
-            std::condition_variable empty;
-        };
-
-        std::unique_ptr<WNMutex> mutex;
-    };
     struct PipelineT {
         PipelineT(orc::ObjectLayer &BaseLayer, TargetMachine &TM, int optlevel);
         CompileLayerT CompileLayer;
@@ -426,17 +292,18 @@ private:
 
     JITDebugInfoRegistry DebugRegistry;
 
-    //Map and inc are guarded by RLST_mutex
-    std::mutex RLST_mutex{};
-    int RLST_inc = 0;
-    DenseMap<void*, std::string> ReverseLocalSymbolTable;
+    struct ReverseLocalSymbolsT {
+        int Unique;
+        DenseMap<void*, std::string> Table;
+    };
+    jl_cc::Locked<ReverseLocalSymbolsT> ReverseLocalSymbols;
 
     //Compilation streams
     jl_locked_stream dump_emitted_mi_name_stream;
     jl_locked_stream dump_compiles_stream;
     jl_locked_stream dump_llvm_opt_stream;
 
-    ResourcePool<orc::ThreadSafeContext, 0, std::queue<orc::ThreadSafeContext>> ContextPool;
+    jl_cc::ResourcePool<orc::ThreadSafeContext, 0, std::queue<orc::ThreadSafeContext>> ContextPool;
 
 #ifndef JL_USE_JITLINK
     const std::shared_ptr<RTDyldMemoryManager> MemMgr;
